Move HLS example to chipyard/example
