+++ generators/chipyard/src/main/scala/DigitalTop.scala
@@ -40,7 +40,6 @@ class DigitalTop(implicit p: Parameters) extends ChipyardSystem
   with fftgenerator.CanHavePeripheryFFT // Enables optionally having an MMIO-based FFT block
   with constellation.soc.CanHaveGlobalNoC // Support instantiating a global NoC interconnect
   with rerocc.CanHaveReRoCCTiles // Support tiles that instantiate rerocc-attached accelerators
-  with hlsaccel.CanHavePeripheryHLSAccel
 {
   override lazy val module = new DigitalTopModule(this)
 }
