+++ generators/chipyard/src/main/scala/example/GCD.scala
@@ -1,5 +1,7 @@
 package chipyard.example
 
+import sys.process._
+
 import chisel3._
 import chisel3.util._
 import chisel3.experimental.{IntParam, BaseModule}
@@ -17,7 +19,8 @@ case class GCDParams(
   address: BigInt = 0x4000,
   width: Int = 32,
   useAXI4: Boolean = false,
-  useBlackBox: Boolean = true)
+  useBlackBox: Boolean = true,
+  useHLS: Boolean = false)
 // DOC include end: GCD params
 
 // DOC include start: GCD key
@@ -35,6 +38,18 @@ class GCDIO(val w: Int) extends Bundle {
   val output_valid = Output(Bool())
   val gcd = Output(UInt(w.W))
   val busy = Output(Bool())
+}
+
+class HLSGCDAccelIO(val w: Int) extends Bundle {
+  val ap_clk = Input(Clock())
+  val ap_rst = Input(Reset())
+  val ap_start = Input(Bool())
+  val ap_done = Output(Bool())
+  val ap_idle = Output(Bool())
+  val ap_ready = Output(Bool())
+  val x = Input(UInt(w.W))
+  val y = Input(UInt(w.W))
+  val ap_return = Output(UInt(w.W))
 }
 
 class GCDTopIO extends Bundle {
@@ -87,6 +102,23 @@ class GCDMMIOChiselModule(val w: Int) extends Module {
   io.busy := state =/= s_idle
 }
 // DOC include end: GCD chisel
+
+// DOC include start: HLS blackbox
+class HLSGCDAccelBlackBox(val w: Int) extends BlackBox with HasBlackBoxPath {
+  val io = IO(new HLSGCDAccelIO(w))
+
+  val chipyardDir = System.getProperty("user.dir")
+  val hlsDir = s"$chipyardDir/generators/chipyard"
+  
+  // Run HLS command
+  val make = s"make -C ${hlsDir}/src/main/resources/hls default"
+  require (make.! == 0, "Failed to run HLS")
+
+  // Add each vlog file
+  addPath(s"$hlsDir/src/main/resources/vsrc/HLSGCDAccelBlackBox.v")
+  addPath(s"$hlsDir/src/main/resources/vsrc/HLSGCDAccelBlackBox_flow_control_loop_pipe.v")
+}
+// DOC include end: HLS blackbox
 
 // DOC include start: GCD router
 class GCDTL(params: GCDParams, beatBytes: Int)(implicit p: Parameters) extends ClockSinkDomain(ClockSinkParameters())(p) {
@@ -190,6 +222,64 @@ class GCDAXI4(params: GCDParams, beatBytes: Int)(implicit p: Parameters) extends
 }
 // DOC include end: GCD router
 
+class HLSGCDAccel(params: GCDParams, beatBytes: Int)(implicit p: Parameters) extends ClockSinkDomain(ClockSinkParameters())(p) {
+  val device = new SimpleDevice("hlsgcdaccel", Seq("ucbbar,hlsgcdaccel")) 
+  val node = TLRegisterNode(Seq(AddressSet(params.address, 4096-1)), device, "reg/control", beatBytes=beatBytes)
+
+  override lazy val module = new HLSGCDAccelImpl
+  class HLSGCDAccelImpl extends Impl with HasGCDTopIO {
+    val io = IO(new GCDTopIO)
+    withClockAndReset(clock, reset) {
+      val x = Reg(UInt(params.width.W))
+      val y = Wire(new DecoupledIO(UInt(params.width.W)))
+      val y_reg = Reg(UInt(params.width.W))
+      val gcd = Wire(new DecoupledIO(UInt(params.width.W)))
+      val gcd_reg = Reg(UInt(params.width.W))
+      val status = Wire(UInt(2.W))
+
+      val impl = Module(new HLSGCDAccelBlackBox(params.width))
+
+      impl.io.ap_clk := clock
+      impl.io.ap_rst := reset
+
+      val s_idle :: s_busy :: Nil = Enum(2)
+      val state = RegInit(s_idle)
+      val result_valid = RegInit(false.B)
+      when (state === s_idle && y.valid) { 
+        state := s_busy
+        result_valid := false.B
+        y_reg := y.bits 
+      } .elsewhen (state === s_busy && impl.io.ap_done) {
+        state := s_idle
+        result_valid := true.B
+        gcd_reg := impl.io.ap_return
+      }
+
+      impl.io.ap_start := state === s_busy
+
+      gcd.valid := result_valid
+      status := Cat(impl.io.ap_idle, result_valid)
+      
+      impl.io.x := x
+      impl.io.y := y_reg
+      y.ready := impl.io.ap_idle
+      gcd.bits := gcd_reg
+
+      io.gcd_busy := !impl.io.ap_idle
+
+      node.regmap(
+        0x00 -> Seq(
+          RegField.r(2, status)), // a read-only register capturing current status
+        0x04 -> Seq(
+          RegField.w(params.width, x)), // a plain, write-only register
+        0x08 -> Seq(
+          RegField.w(params.width, y)), // write-only, y.valid is set on write
+        0x0C -> Seq(
+          RegField.r(params.width, gcd))) // read-only, gcd.ready is set on read
+    }
+  }
+}
+
 // DOC include start: GCD lazy trait
 trait CanHavePeripheryGCD { this: BaseSubsystem =>
   private val portName = "gcd"
@@ -209,6 +299,15 @@ trait CanHavePeripheryGCD { this: BaseSubsystem =>
           // toVariableWidthSlave doesn't use holdFirstDeny, which TLToAXI4() needsx
           TLFragmenter(pbus.beatBytes, pbus.blockBytes, holdFirstDeny = true) := _
         }
+        gcd
+      } else if (params.useHLS) {
+        // val gcd = LazyModule( 
+        //   if (params.useHLS) new HLSGCDAccel(params, pbus.beatBytes)(p) 
+        //   else new GCDTL(params, pbus.beatBytes)(p)
+        // )
+        val gcd = LazyModule(new HLSGCDAccel(params, pbus.beatBytes)(p))
+        gcd.clockNode := pbus.fixedClockNode
+        pbus.coupleTo(portName) { gcd.node := TLFragmenter(pbus.beatBytes, pbus.blockBytes) := _ }
         gcd
       } else {
         val gcd = LazyModule(new GCDTL(params, pbus.beatBytes)(p))
@@ -233,3 +332,8 @@ class WithGCD(useAXI4: Boolean = false, useBlackBox: Boolean = false) extends Co
   case GCDKey => Some(GCDParams(useAXI4 = useAXI4, useBlackBox = useBlackBox))
 })
 // DOC include end: GCD config fragment
+
+// useHLS cannot be used with useAXI4 and useBlackBox
+class WithHLSGCD extends Config((site, here, up) => {
+  case GCDKey => Some(GCDParams(useAXI4 = false, useBlackBox = false, useHLS = true))
+})
